fix(mqtt): integration-tag lint (lostcancel + unused dispatchAck doubles) [V11-4 PR-3]
The prior commit's SessionRecovery rewrite + dispatchAck test doubles broke the
integration-tag build/lint (CI: make verify + integration-test red), which the
unit-only local checks did not catch:

- integration_test.go SessionRecovery: the derived subCtx1/subCancel1 was only
  canceled on the timeout path → govet lostcancel (context leak) on the happy
  path. sub1.Subscribe now runs under ctx1 directly (canceled at phase-1 teardown
  to drop the manager uncleanly + retain the persistent session), removing the
  leak.
- subscriber_testhelpers_test.go (no build tag) held dispatchAck white-box
  doubles (subAckRecorder/fakeAckConn/newDispatchAckSubscriber/dispatchAckEntry/
  ackHandler/testingT) whose only callers are unit-only (subscriber_test.go is
  !integration), so they reported  under -tags=integration. Moved them to
  subscriber_dispatchack_test.go (//go:build !integration); recordingSettlement /
  recordingSubCollector stay no-tag (used by both unit + integration).

Verified locally with Docker: all 7 TestIntegration_* PASS (incl.
Subscriber_SessionRecovery 11.77s, Disposition3State ack/requeue/reject,
ClientIDConflict); lint clean under both default and --build-tags=integration;
unit -race green.
